added inventory for Artifacts. Character can add new artifacts to inventory
Added scrollable inventory with item name and description for any artifact character may find. Just create a new BP child from BP_Interactable and put it in the world.
The interactable artifacts are also highlighted during the scan effect (need to check custom rendering, value 1 or 2 in the details panel for highlighting)